# sdk/openapi.json
{
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "type": "object"
          }
        },
        "type": "object"
      },
...
      "Registry": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "RegistryList": {
        "items": {
          "$ref": "#/components/schemas/Registry"
        },
        "type": "array"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "A tenant API key, or an OAuth access token.",
        "scheme": "bearer",
        "type": "http"
      },
      "oauth2": {
        "description": "An OAuth 2.1 access token of this tenant.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Isolated Linux microVMs on one host, created from OCI images and forked into copies.",
    "title": "Kiln",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/registries": {
      "get": {
        "operationId": "listRegistries",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistryList"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "List the registry credentials of this tenant. The tokens are never returned."
      },
      "post": {
        "operationId": "putRegistry",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "host": {
                    "type": "string"
                  },
                  "token": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                },
                "required": [
                  "host",
                  "username",
                  "token"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Registry"
                }
              }
            },
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Store the credential a template build uses to pull a private image for this tenant."
      }
    },
    "/v1/registries/{host}": {
      "delete": {
        "operationId": "deleteRegistry",
        "parameters": [
          {
            "description": "Registry hostname.",
            "in": "path",
            "name": "host",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The call succeeded."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The request is invalid. The message names the field."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The credential is missing, expired or revoked."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The row does not exist, or it belongs to another tenant."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "The call conflicts with the state, for example a template that still has sandboxes."
          },
          "507": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "A cap of this tenant is reached. The message names the cap."
          }
        },
        "summary": "Remove one registry credential."
      }
    },
...
  },
  "security": [
    {
      "apiKey": []
    },
    {
      "oauth2": []
    }
  ],
  "servers": [
    {
      "url": "https://api.example.com"
    }
  ]
}